+++ X86Disassembler/X86/Handlers/Jump/JmpRel8Handler.cs
@@ -49,7 +49,7 @@ public class JmpRel8Handler : InstructionHandler
         ulong targetAddress = instruction.Address + 2UL + (uint)offset;
         
         // Create the target address operand
-        var targetOperand = OperandFactory.CreateRelativeOffsetOperand(targetAddress, 8);
+        var targetOperand = OperandFactory.CreateRelativeOffsetOperand((uint)targetAddress, 8);
         
         // Set the structured operands
         instruction.StructuredOperands =
