+++ X86Disassembler/X86/Handlers/Sub/SubImmFromRm16SignExtendedHandler.cs
@@ -78,7 +78,7 @@ public class SubImmFromRm16SignExtendedHandler : InstructionHandler
         short imm16 = (sbyte)Decoder.ReadByte();
 
         // Create the source immediate operand with the sign-extended value
-        var sourceOperand = OperandFactory.CreateImmediateOperand(imm16, 16);
+        var sourceOperand = OperandFactory.CreateImmediateOperand((ushort)imm16, 16);
         
         // Set the structured operands
         instruction.StructuredOperands =
